Crate cucumber

Reexports

pub use runner::{CommandRunner, WorldRunner};

pub use definitions::registration::CucumberRegistrar;

pub use state::{Cucumber, SendableStep};

pub use server::Server;

pub use event::request::InvokeArgument;

Modules

cucumber_regex
Helpers for regular expressions

definitions
Business logic for step registration and invoke argument destructuring

event
External facing interface for events

runner
Coordinator logic between server and state

server
External facing interface to other Gherkin implementations

state
Low level location of step functions and matcher logic

Macros

Given!
Add a Given step to a CucumberRegistrar

Then!
Add a Then step to a CucumberRegistrar

When!
Add a When step to a CucumberRegistrar

try_destructure!
Destructure a vector of InvokeArgument into a tuple of values, or a bad InvokeResponse, similar to normal try!

Structs

CucumberConfig
Representation of the cucumber server and client configuration

Functions

create_config
Configure the Cucumber server and Ruby client

ruby_command
Build a command to execute the Ruby Cucumber Server. Takes a list of extra command line arguments to the server NOTE: This command also passes the command line arguments passed to this executable.