Struct cucumber::runner::WorldRunner

pub struct WorldRunner<World> {
    // some fields omitted
}

The step runner for Cucumber state

The runner stands in for the Cucumber instance and provides an interface for Request events to be translated into state changes and step invocations, along with a Response. These are typically supplied by a running Server, but may be supplied by a native Gherkin implementation later.

Typically this struct will only be instantiated by the user, and then passed to a Server to maintain.

Methods

impl<World> WorldRunner<World>

fn new(world: World) -> WorldRunner<World>