Struct cucumber::server::ServerHandle

pub struct ServerHandle {
    // some fields omitted
}

An "unplugged" return value from server start, for compiler reasons

The "correct" return value for Server#start is this server handle. However, it yields a weird linker error on stable when methods are invoked on it when the crate is imported (so not on local tests). Therefore, we're currently returning the components of this structure as a tuple for now.

See build: https://travis-ci.org/acmcarther/cucumber/jobs/116256537 Example Error /home/travis/build/acmcarther/cucumber-rs/examples/calculator/features/cuke. rs:31: undefined reference to `server::ServerHandle::wait::hd35f2fdfe2f62e1dyv

Methods

impl ServerHandle

fn stop(&mut self)

fn wait(self)