Struct cucumber::CucumberConfig

#[must_use = "CucumberConfig has to be consumed by start method!"]
pub struct CucumberConfig<'a, W: Send + 'static> {
    // some fields omitted
}

Representation of the cucumber server and client configuration

A new configuration can be generated by calling the function create_config

Methods

impl<'a, W: Send + 'static> CucumberConfig<'a, W>

fn address(self, address: &'static str) -> CucumberConfig<'a, W>

Adds a custom ip and port, that will replace the default of localhost:7878

fn registrar_fns(self, registrars: &'a [&Fn(&mut CucumberRegistrar<W>)]) -> CucumberConfig<'a, W>

Adds a slice of registrar functions.

fn registrar_fn(self, registrar: &'a Fn(&mut CucumberRegistrar<W>)) -> CucumberConfig<'a, W>

Adds a single registrar functions

fn args(self, args: &'a [&'static str]) -> CucumberConfig<'a, W>

Adds a slice of arguments that is passed to the ruby client

fn arg(self, arg: &'static str) -> CucumberConfig<'a, W>

Adds a single argument that is passed to the ruby client

fn start(self)

Starts Cucumber and the ruby client using the defined settings.